+++ tests/src/CommandsTestCase.php
@@ -9,6 +9,7 @@
 namespace Gecche\Multidomain\Tests;
 
 use Gecche\Multidomain\Foundation\Application;
+use Gecche\Multidomain\Foundation\Configuration\ApplicationBuilder;
 use Gecche\Multidomain\Foundation\Providers\DomainConsoleServiceProvider;
 use Gecche\Multidomain\Tests\Http\Kernel as HttpKernel;
 use Illuminate\Filesystem\Filesystem;
@@ -16,6 +17,10 @@ use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Config;
 use Gecche\Multidomain\Tests\Console\Kernel as ConsoleKernel;
+
+use Illuminate\Contracts\Console\Kernel as ConsoleKernelContract;
+use Illuminate\Contracts\Http\Kernel as HttpKernelContract;
+
 
 class CommandsTestCase extends \Orchestra\Testbench\TestCase
 {
@@ -75,22 +80,36 @@ class CommandsTestCase extends \Orchestra\Testbench\TestCase
      */
     protected function resolveApplicationConsoleKernel($app)
     {
-        $app->singleton('Illuminate\Contracts\Console\Kernel', ConsoleKernel::class);
+        $app->singleton(ConsoleKernelContract::class, ConsoleKernel::class);
     }
 
     protected function resolveApplicationHttpKernel($app)
     {
-        $app->singleton('Illuminate\Contracts\Http\Kernel', HttpKernel::class);
+        $app->singleton(HttpKernelContract::class, HttpKernel::class);
     }
 
     protected function resolveApplication()
     {
-        return tap(new Application($this->getBasePath()), function ($app) {
-            $app->bind(
-                'Illuminate\Foundation\Bootstrap\LoadConfiguration',
-                'Orchestra\Testbench\Bootstrap\LoadConfiguration'
-            );
-        });
+        static::$cacheApplicationBootstrapFile ??= $this->getApplicationBootstrapFile('app.php');
+
+        if (\is_string(static::$cacheApplicationBootstrapFile)) {
+            $APP_BASE_PATH = $this->getBasePath();
+
+            return require static::$cacheApplicationBootstrapFile;
+        }
+
+        return $this->resolveGeccheApplication();
+    }
+
+    final protected function resolveGeccheApplication()
+    {
+        return (new ApplicationBuilder(new Application($this->getBasePath())))
+            ->withProviders()
+            ->withMiddleware(static function ($middleware) {
+                //
+            })
+            ->withCommands()
+            ->create();
     }
 
     /**
@@ -101,7 +120,7 @@ class CommandsTestCase extends \Orchestra\Testbench\TestCase
      */
     protected function getEnvironmentSetUp($app)
     {
-
+        // set up database configuration
     }
 
     /**
@@ -114,6 +133,7 @@ class CommandsTestCase extends \Orchestra\Testbench\TestCase
         return [
             TestServiceProvider::class,
             DomainConsoleServiceProvider::class,
+//            ServiceProvider::class,
         ];
     }